--- src/test/java/com/CRM_project/pages/Test23.java
@@ -1,9 +1,0 @@
-package com.CRM_project.pages;
-
-public class Test23 {
-
-    public static void main(String[] args) {
-
-        System.out.println("Hello hello");
-    }
-}
