--- src/test/java/com/CRM_project/pages/TestPage.java
@@ -1,12 +1,0 @@
-package com.CRM_project.pages;
-
-public class TestPage {
-
-    public static void main(String[] args) {
-
-        System.out.println("Hello");
-
-    }
-
-
-}
